% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{compute_group_timeline}
\alias{compute_group_timeline}
\title{Compute Group Function for StatTimeline}
\usage{
compute_group_timeline(data, scales)
}
\arguments{
\item{data}{The data based on required and optional aesthetics of geom}

\item{scales}{The scales of geom}
}
\value{
The function returns the data as prepared for the geom's draw panel function
}
\description{
The function filters the earthquake data rows such that the date of earthquake occurrence
falls between the provided minimum and maximum dates for the time line
}
\examples{
\dontrun{
Internal function called by Stat when geom is called
}
}
