% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{compute_group_timelinelabel}
\alias{compute_group_timelinelabel}
\title{Compute Group Function for StatTimelineLabel}
\usage{
compute_group_timelinelabel(data, scales, n_max = 4)
}
\arguments{
\item{data}{The data based on required and optional aesthetics of geom}

\item{scales}{The scales of geom}

\item{n_max}{Number of largest earthquakes (by magnitude), for which labels should be displayed}
}
\value{
The function returns the data which can be used by geom's draw panel function
}
\description{
The function filters the earthquake data rows such that the magnitude of earthquake
is greater than or equal to the maximum
}
\examples{
\dontrun{
Internal function called by Stat when geom is called
}
}
