% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{draw_panel_timeline}
\alias{draw_panel_timeline}
\title{Draw panel function for ggproto to extend GeomPoint}
\usage{
draw_panel_timeline(data, panel_scales, coord)
}
\arguments{
\item{data}{The data that is passed after computation from compute_group function of StatTimeline}

\item{panel_scales}{The panel scales based on which data is transformed via coord}

\item{coord}{The coordinates of the ggmap. which has transform function}
}
\value{
PointsGrob object to be drawn on ggplot2 layer
}
\description{
This function is called when the custom geom is been drawn on its layer. It returns the
PointsGrob, which contains all the points along with its attributes of shape & size.
}
\examples{
\dontrun{
Internal function called by ggproto when this extended GeomPoint is called
}
}
