% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{draw_panel_timelinelabel}
\alias{draw_panel_timelinelabel}
\title{Draw panel function for ggproto to extend Geom class}
\usage{
draw_panel_timelinelabel(data, panel_scales, coord)
}
\arguments{
\item{data}{The data that is passed after computation from compute_group function of StatTimelineLabel}

\item{panel_scales}{The panel scales based on which data is transformed via coord}

\item{coord}{The coordinates of the ggmap. which has the transform function}
}
\value{
gtree object with 2 children segment & text to show the earthquakes label
}
\description{
This function is called when the custom geom function is called to draw trees of new
geometric objects on a new layer. It returns the gtree object composed of segmentGrob
and textGrob for the filtered earthquakes that are top largest in magnitudes
}
\examples{
\dontrun{
Internal function called by ggproto when this extended Geom is called
}
}
