% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans the Earthquake raw data}
\usage{
eq_clean_data(df_raw = NULL)
}
\arguments{
\item{df_raw}{The source data as loaded from the raw files}
}
\value{
This function returns the cleaned data with date format based DATE, appropriate
   LOCATION_NAME and other selected fields after been transformed into correct data types
}
\description{
The NOAA earthquake data is available as tab delimited file. The function assumes
that the data is provided in raw form directly after reading the files from I/O.
}
\examples{
eqvis::eq_clean_data()

}
