% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates an HTML label that can be used as the annotation text in the leaflet map.}
\usage{
eq_create_label(df = NULL)
}
\arguments{
\item{df}{The clean data as returned from eq_clean_data}
}
\value{
A vector that consists of HTML labels for each earthquake
}
\description{
This function puts together a character string for each earthquake that will show the
cleaned location (using eq_location_clean), the magnitude (EQ_PRIMARY), and
the total number of deaths (TOTAL_DEATHS). Also, it adds boldface labels for each
("Location: ", "Total deaths: ", and "Magnitude: "). If an earthquake is missing values
for any of these, both the label and the value is skipped for that element of the tag.
}
\examples{
eqvis::eq_load_data() \%>\%
  eqvis::eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  dplyr::select("DATE","LATITUDE","LONGITUDE","LOCATION_NAME","TOTAL_DEATHS","EQ_PRIMARY","popup_text")
}
