% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_load_data}
\alias{eq_load_data}
\title{Loads the Data from delimited file}
\usage{
eq_load_data(filename = "")
}
\arguments{
\item{filename}{The location of the file with file name}
}
\value{
Data Frame with raw data
}
\description{
Loads the Data from delimited file
}
\examples{
eqvis::eq_load_data()

}
