% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans the LOCATION_NAME column}
\usage{
eq_location_clean(df_raw = NULL)
}
\arguments{
\item{df_raw}{The source data as loaded from the raw files}
}
\value{
The source data having cleaned LOCATION_NAME & factored COUNTRY name
}
\description{
This function should be applied to the raw data to produce a cleaned up version
of the LOCATION_NAME column.It stripes out the country name (including the colon) and
converts the remaining location names into title case.
}
\examples{
eqvis::eq_location_clean()

}
