% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_map}
\alias{eq_map}
\title{Uses leaflet functions to map the epicenters (LATITUDE/LONGITUDE) and also to annotate
each point with a pop up window.}
\usage{
eq_map(df_map = NULL, annot_col = "DATE")
}
\arguments{
\item{df_map}{The cleaned earthquake data set, having the column with name matching the value of annot_col}

\item{annot_col}{The name of the column that contains HTML text to be displayed in popup}
}
\value{
A leaflet map widget showing the epicenters along with popup text
}
\description{
Uses leaflet functions to map the epicenters (LATITUDE/LONGITUDE) and also to annotate
each point with a pop up window.
}
\examples{
eqvis::eq_load_data() \%>\%
  eqvis::eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eqvis::eq_create_label(.)) \%>\%
  eqvis::eq_map(annot_col = "popup_text")
}
