% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_map_with_datelabels}
\alias{eq_map_with_datelabels}
\title{Map Earthquake Locations showing the magnitude and dates of occurrences}
\usage{
eq_map_with_datelabels(countries = c("MEXICO"), minyear = 2000)
}
\arguments{
\item{countries}{Character vectors of country names}

\item{minyear}{Integer representing minimum year}
}
\value{
Map showing Earthquake locations along with popup text
}
\description{
Map Earthquake Locations showing the magnitude and dates of occurrences
}
\examples{
eq_map_with_datelabels()
}
