% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_map_with_htmllabels}
\alias{eq_map_with_htmllabels}
\title{Map Earthquake Locations along with popup for key data formatted with HTML}
\usage{
eq_map_with_htmllabels(countries = c("MEXICO"), minyear = 2000)
}
\arguments{
\item{countries}{Character vectors of country names}

\item{minyear}{Integer representing minimum year}
}
\value{
Map showing Earthquakes along with popup for Location, Magnitude & Total Deaths
}
\description{
Map Earthquake Locations along with popup for key data formatted with HTML
}
\examples{
eq_map_with_htmllabels()
}
