% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{eq_timeline}
\alias{eq_timeline}
\title{Plots the Timelines and the popup text for earthquakes largest in size}
\usage{
eq_timeline(
  countries = c("TURKEY", "CHINA"),
  minymd = 20000101,
  maxymd = 20220101,
  mineqmag = 0
)
}
\arguments{
\item{countries}{Character vectors of country names}

\item{minymd}{Integer Year Month Day of timeline start}

\item{maxymd}{Integer Year Month Day of timeline end}

\item{mineqmag}{Numeric minimum magnitude of earthquakes to be ploted}
}
\value{
Visualization in form of gemetric objects plotted on 2d chart
}
\description{
Plots the Timelines and the popup text for earthquakes largest in size
}
\examples{
eqvis::eq_timeline()
}
