% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{geom_timeline}
\alias{geom_timeline}
\title{The main function for GeomPoint extension for Earthquake Timelines visualization}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = TRUE,
  inherit.aes = TRUE,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{mapping}{The mapping of aesthetics with data columns}

\item{data}{The data that needs to be visualized}

\item{stat}{Has default value as identity}

\item{position}{Has default value as identity}

\item{show.legend}{TRUE indicates legend has to be shown}

\item{inherit.aes}{TRUE indicates that aesthetics are inherited from base layer}

\item{na.rm}{FALSE indicates that the NA values needs to be displayed}

\item{...}{Any other parameters}
}
\value{

}
\description{
The main function for GeomPoint extension for Earthquake Timelines visualization
}
\examples{
\dontrun{
eqvis::eq_load_data('../capstone/data/earthquakes.tsv.gz') \%>\%
eqvis::eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("TURKEY","CHINA") & lubridate::year(DATE) >= 2000 & lubridate::year(DATE) < 2022 & EQ_MAG_MW>=0) \%>\%
ggplot2::ggplot(mapping = aes(x=DATE, y=COUNTRY)) +
eqvis::geom_timeline(ggplot2::aes(xmin=lubridate::ymd(20000101),
                          xmax=lubridate::ymd(20220101),
                          size=EQ_MAG_MW,
                          color=TOTAL_DEATHS,
                          shape=COUNTRY,
                          alpha=0.6
))
}
}
