% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{The main function for Geom extension that adds Earthquake labels on Timelines}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = FALSE,
  inherit.aes = TRUE,
  na.rm = FALSE,
  n_max = 4,
  ...
)
}
\arguments{
\item{mapping}{The mapping of aesthetics with data columns}

\item{data}{The data that needs to be visualized}

\item{stat}{Has default value as identity}

\item{position}{Has default value as identity}

\item{show.legend}{TRUE indicates legend has to be shown}

\item{inherit.aes}{TRUE indicates that aesthetics are inherited from base layer}

\item{na.rm}{FALSE indicates that the NA values needs to be displayed}

\item{n_max}{The total number of earthquakes (largest by magnitude) that needs to be labeled}

\item{...}{Any other parameters}
}
\value{
New ggplot2 Layer having values that refer to StatTimelineLabel & GeomTimelineLabel
}
\description{
The main function for Geom extension that adds Earthquake labels on Timelines
}
\examples{
\dontrun{
eqvis::eq_load_data('../capstone/data/earthquakes.tsv.gz') \%>\%
eqvis::eq_clean_data() \%>\%
dplyr::filter(COUNTRY \%in\% c("TURKEY","CHINA") & lubridate::year(DATE) >= 2000 & lubridate::year(DATE) < 2022 & EQ_MAG_MW>=0) \%>\%
ggplot2::ggplot(mapping = aes(x=DATE, y=COUNTRY)) +
eqvis::geom_timeline(ggplot2::aes(xmin=lubridate::ymd(20000101),
                          xmax=lubridate::ymd(20220101),
                          size=EQ_MAG_MW,
                          color=TOTAL_DEATHS,
                          shape=COUNTRY,
                          alpha=0.6
)) +
eqvis::geom_timeline_label(ggplot2::aes(xmin=lubridate::ymd(20000101),
                          xmax=lubridate::ymd(20220101),
                          size=EQ_PRIMARY,
                          color=TOTAL_DEATHS,
                          shape=COUNTRY,
                          alpha=0.6,
                          label=LOCATION_NAME
                          ),
                   n_max=4
                   )
}
}
