% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{make_filename}
\alias{make_filename}
\title{Returns the package based filename}
\usage{
make_filename()
}
\value{
This function returns the string filename which if exists contain
  the data for the earthquakes occurrences
}
\description{
The NOAA earthquakes data is available as an archived file having tab
delimed data. The function assumes the location of file within the package
and returns the filename for loading raw data.
}
\examples{
\dontrun{make_filename()}

# No export
}
