% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\name{stat_timeline_label}
\alias{stat_timeline_label}
\title{Uses StatTimelineLabel as stat for generating new ggplot2 Layer}
\usage{
stat_timeline_label(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = FALSE,
  n_max = 4,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and
inherit.aes = TRUE (the default), it is combined with the default mapping at the top
level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If NULL, the default, the data is inherited from the plot data as specified
in the call to ggplot().
A data.frame, or other object, will override the plot data. All objects will be fortified
to produce a data frame. See fortify() for which variables will be created.
A function will be called with a single argument, the plot data. The return value must be
a data.frame, and will be used as the layer data. A function can be created from a
formula (e.g. ~ head(.x, 10)).}

\item{geom}{The geometric object to use display the data on ggplot layer}

\item{position}{Position adjustment, either as a string, or the result of a call to a
position adjustment function, as required for ggplot layer}

\item{show.legend}{Logical indicator to show the Key Legend on ggplot layer}

\item{inherit.aes}{Logical indicator whether aesthetics to be inherited from prev layer}

\item{na.rm}{Logical indicator whether NA to be removed}

\item{n_max}{The top n_max largest earthquakes (by magnitude) that should be selected}

\item{...}{Additional parameters to the stat layer}
}
\value{
Returns the ggplot2::layer Object
}
\description{
Each ggplot2 layer provides stat parameter, that can perform required computation
on data in order to provide the same to next layer. This function passes StatTimelineLabel,
the parameterized ggproto function, which uses compute_group_timelinelabel function for
earthquake data that selects the top n_max largest earthquakes (by magnitude).
}
\examples{
\dontrun{
Function not used but defined for compatibility and future use
}
}
