% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquakes.R
\docType{data}
\name{StatTimelineLabel}
\alias{StatTimelineLabel}
\title{StatTimelineLabel - A new class that inherits from the Stat and computes top largest Earthquakes for labeling}
\format{
An object of class \code{StatTimelineLabel} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatTimelineLabel
}
\description{
compute_group The function that prepares the data based on which geometric object is drawn
}
\details{
Required Aesthetics:
xmin The MINIMUM date for the earthquake timeline
xmax The MAXIMUM date for the earthquake timeline
colour The border color for point. Default is GREEN
size The size of the point. Default is 2
alpha The transparency value for fill color. Default is 0.6

Default Aesthetics:
y A factor indicating some stratification in which case multiple timelines will be plotted
  for each level of the factor (e.g. country).

Extra Parameters:
na.rm Logical indicator for whether NA values need to be removed or not
n_max Number of total earthquakes, largest (by magnitude), which needs to be labeled
}
\keyword{datasets}
